pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "desktop-agent-mvp"
version = "0.1.0"
description = "Windows desktop automation MVP with planner API and executor loop"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "httpx>=0.27.0",
  "pillow>=10.4.0",
]

[project.optional-dependencies]
ocr = [
  "pytesseract>=0.3.13",
]
desktop = [
  "pyautogui>=0.9.54",
  "pynput>=1.7.7",
]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "packages*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
